+++ src/Accounts/Authentication/Factories/AuthenticationFactory.cs
@@ -573,14 +573,29 @@ namespace Microsoft.Azure.Commands.Common.Authentication.Factories
                     password = password ?? account.GetProperty(AzureAccount.Property.ServicePrincipalSecret)?.ConvertToSecureString();
                     if (password == null)
                     {
-                        password = KeyStore.GetKey(new ServicePrincipalKey(AzureAccount.Property.ServicePrincipalSecret
-                        , account.Id, tenant));
+                        try
+                        {
+                            password = KeyStore.GetKey<SecureString>(new ServicePrincipalKey(AzureAccount.Property.ServicePrincipalSecret
+, account.Id, tenant));
+                        }
+                        catch
+                        {
+                            password = null;
+                        }
+
                     }
                     var certificatePassword = account.GetProperty(AzureAccount.Property.CertificatePassword)?.ConvertToSecureString();
                     if (certificatePassword == null)
                     {
-                        certificatePassword = KeyStore.GetKey(new ServicePrincipalKey(AzureAccount.Property.CertificatePassword
-                        , account.Id, tenant));
+                        try
+                        {
+                            certificatePassword = KeyStore.GetKey<SecureString>(new ServicePrincipalKey(AzureAccount.Property.CertificatePassword
+                            , account.Id, tenant));
+                        }
+                        catch
+                        {
+                            certificatePassword = null;
+                        }
                     }
                     return new ServicePrincipalParameters(tokenCacheProvider, environment, tokenCache, tenant, resourceId, account.Id, account.GetProperty(AzureAccount.Property.CertificateThumbprint), account.GetProperty(AzureAccount.Property.CertificatePath),
                         certificatePassword, password, sendCertificateChain);